src/main/java/edu/yctc/genesis/entity/KnowledgeDO.java
package edu.yctc.genesis.entity;

import java.util.Date;

public class KnowledgeDO {

    /** 主键 */
    private long id;
    /** 知识点描述 */
    private String content;
    /** 预计时长/分钟 */
    private double length;
    /** 记录时间 */
    private Date createTime;
    /** 修改时间 */
    private Date modifyTime;

    public long getId() {
        return id;
    }

    public void setId(long id) {
        this.id = id;
    }

    public String getContent() {
        return content;
    }

    public void setContent(String content) {
        this.content = content;
    }

    public double getLength() {
        return length;
    }

    public void setLength(double length) {
        this.length = length;
    }

    public Date getCreateTime() {
        return createTime;
    }

    public void setCreateTime(Date createTime) {
        this.createTime = createTime;
    }

    public Date getModifyTime() {
        return modifyTime;
    }

    public void setModifyTime(Date modifyTime) {
        this.modifyTime = modifyTime;
    }

    @Override
    public String toString() {
        StringBuilder builder = new StringBuilder();
        builder.append("Knowledge [id=");
        builder.append(id);
        builder.append(", content=");
        builder.append(content);
        builder.append(", length=");
        builder.append(length);
        builder.append(", createTime=");
        builder.append(createTime);
        builder.append(", modifyTime=");
        builder.append(modifyTime);
        builder.append("]");
        return builder.toString();
    }

}
